+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 - Added Python 3.4 support.
 - Added :meth:`Connection.wait_closed`.
 - Added ``auto_link`` parameter to :func:`connect`, :func:`start_server` and :func:`run_app`.
+- Added :attr:`Connection.hostname`.
 
 
 2.1.0 - 04/11/2016